+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 0.13.2 -- Apr 23, 2021
 * Update `setuptools_certificate` dependency to new name `otumat`
 * Bugfix - Explicit calls to `dj.Connection` throw error due to missing `host_input` (#895) PR #907
+* Bugfix - Correct count of deleted items. (#897) PR #912
 
 ### 0.13.1 -- Apr 16, 2021
 * Add `None` as an alias for `IS NULL` comparison in `dict` restrictions (#824) PR #893